requirements.txt
streamlit>=1.35.0
openai-whisper
openai
torch
transformers
gTTS
requests
python-dotenv
pytest
